# internal/api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Arduino-App-Cli
  version: 0.1.0
servers:
- description: local server
  url: http://localhost:8800
...
paths:
...
  /v1/apps/{id}/start:
    post:
      description: Start the application and handles all the operation to start any
        dependecies. If the app contains a sketch it also flash it in the micro.
      operationId: startApp
      parameters:
      - description: Return a verbose output. Default is false.
        in: query
        name: verbose
        schema:
          description: Return a verbose output. Default is false.
          type: boolean
      - description: application identifier.
        in: path
        name: id
        required: true
        schema:
          description: application identifier.
          type: string
      responses:
        "200":
          content:
            text/event-stream:
              schema:
                type: string
          description: |
            A stream of Server-Sent Events (SSE) that notifies the progress.
            The client will receive events formatted as follows:

            **Event 'progress'**:
            Contains a JSON object with the percentage of completion.
            'event: progress'
            'data: {"progress":0.25}'

            **Event 'message'**:
            Contains a JSON object with an informational message.
            'event: message'
            'data: {"message":"Starting container..."}'

            **Event 'error'**:
            Contains a JSON object with the details of an error.
            'event: error'
            'data: {"code":"INTERNAL_SERVER_ERROR","message":"An error occurred during operation"}'
        "412":
          $ref: '#/components/responses/PreconditionFailed'
        "500":
          $ref: '#/components/responses/InternalServerError'
      summary: Start an existing app/example
      tags:
      - Application
...
components:
  responses:
...
    InternalServerError:
      content:
        application/json:
          example:
            details: An unexpected error occurred.
          schema:
            $ref: '#/components/schemas/ErrorResponse'
      description: Internal Server Error
...
    PreconditionFailed:
      content:
        application/json:
          example:
            details: The request is invalid.
          schema:
            $ref: '#/components/schemas/ErrorResponse'
      description: Precondition Failed
...
  schemas:
...
    ErrorResponse:
      properties:
        code:
          type: integer
        message:
          type: string
      type: object
...